pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dailywriting"
version = "1.0.0"
description = "A macOS app for daily writing practice with AI-powered prompts"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "SL-blue" }
]
dependencies = [
    "PyQt6>=6.10.0",
    "google-genai>=1.51.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-qt>=4.4.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
dailywriting = "main:main"

[project.gui-scripts]
DailyWriting = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["core", "ui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.coverage.run]
source = ["core", "ui"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
